test: add rudimentary unit test for cnv_json
The main reason for adding this at the moment is that after removing the
dummy script, there are no unit tests to collect, resulting in pytest
giving back non-zero exit status. This is apparently the expected
behaviour but it breaks our CI. With this there is at least a couple of
tests to collect.

These tests should exist anyways, and more tests should be added.