fix: don't generate copy rules if used as module
If a Hydra Genetics workflow using the copy-rules for the final results
imports a module that also does this, there will be ambiguous rules since
both read from the same output file specification, at least if they use
the same config. This commit addresses this by checking if there are any
modules loaded. If that is the case, then it is not likely that it is
being run as a standalone workflow, and the copy-rules should not be
generated.