docs: detail how to override rules and keep some defaults
When overriding rules it is not possible to only overload parts of a
section, e.g. only a part of the input. This became a problem after I
changed the way the template is being fetched. I did however find a
workaround which involves making use of the workflow object to fetch
defaults that should not be overloaded.

I haven't tried building this into the logic since that would likely
require a change to the Snakemake codebase. There is an [issue
about this](https://github.com/snakemake/snakemake/issues/928) in
the Snakemake repo, but it hasn't recieved any attention for the
past couple of years, so I assume this is not prioritised.