feat: improve plotting performance (#42)
* Larger test dataset for performance testing

* Sort plotting elements according to position

* First attempt at smarter chromosome plotting

* Only consider data within the domain

* Improved sliding window logic

This fixes a couple of issues:

1. Early returns result in no data for some ranges
2. Early returns result in data that extends beyond the view

* Use the sliding window approach for the genome view

* Add window size parameter

Also use a smaller window size for the genome view to compensate for the
shorter chromosomes. This should make the overall number of data points
in the two views more similar, but ideally this should be adjusted
according to the number of chromosomes in the view.

* Update test to use new test data

* Pick a random point in each window

* Better summary of windows, including mean and stdev

At the moment this only includes the log2 ratios for the chromosome
plot. Everything else is more or less broken at the moment.

* Fix transition bug when changing chromosomes quickly

* Implement window summary for genome ratio panel

Now it's just the VAF parts that need to be implemented.

* Summarise VAF in the chromosome plot

This uses a line to represent the mean instead of a point as in the case
of the log2 ratio. I will change the log2 ratio to also use a line.

* Summarise VAF in the genome plot

* Fix exit transition in genome plot

* Fix update bug when zoomed in to individual points

* Represent mean with a line instead of a point

* Refactor window summary functions

* Check that windows are not empty

* Fix enter bug for individual points in genome plot

* Ability to handle multiple AF values for a variant

* Only summarise if more than 300 points per panel

* Get text dimensions more efficiently

* Plot windows with constant width

* Flatten VAF arrays

* Add force-summarise to window function

* Summarise all or none of the VAFs in the genome view

* Summarise all or none of the ratios in the genome view

* Only flatten VAF array if it is actually nested

* Ignore duplicate annotations

* Add control for showing all data

* Fix off-by-one error

This was not an obvious error when looking at real data that was
relatively sparse. However, with the pretty dense test data, this effect
became apparent where windows became misaligned the further right in the
plot you looked.

* Only consider data points that would be visible

* Remove unnecessary files

* Fix module dry-run

* Better segment colour contrast

* Apptainer might need the source cache directory

When overloading a rule that uses the source cache when at the same time
running the pipeline using Apptainer, the user has to make sure that
this directory is mounted.