Address filtering bug
The reason for the weird behaviour was that if multiple CNVs in cnvkit
overlapped a single GATK segment, and the first segment group did not
pass the filter, the GATK segment would be set to not pass the filter,
and then never be evaluated properly again since it had already been
considered. This fix addresses this by keeping track of the filter state
and updating it if it turns out that it would pass in comparisons down
the road. Instead of adding the CNVs to the final results in the loop,
all CNVs for a chromosome are added only when the whole chromosome has
been checked.

I should add some more tests though to make sure that this doesn't break
in other cases.