fix: table filtering issue (#75) (#76)
* Refactor to prepare for unit testing

This removes the need to access the snakemake object in any of the
functions. Otherwise this would break outside the context of a pipeline.

* Modify test VCFs to reproduce table filtering issue

I'm not completely sure on the mechanics of this, but I think that this
thing manifests itself (in this particular case) when there are a number
of segments in cnvkit that overlap a single larger segment in GATK, and
the first of the cnvkit segments is COPY_NORMAL.

I have to do some more experimentation to get to the bottom of it.

* Reorganise unit tests under .tests/units

* Add test that catches filtering issue

* Fix generator type hint

typing.Generator is deprecated from python 3.9 in favour of
collections.abc.Generator. The tests on Github run on python 3.8, and I
was testing locally with 3.9.

* Address filtering bug

The reason for the weird behaviour was that if multiple CNVs in cnvkit
overlapped a single GATK segment, and the first segment group did not
pass the filter, the GATK segment would be set to not pass the filter,
and then never be evaluated properly again since it had already been
considered. This fix addresses this by keeping track of the filter state
and updating it if it turns out that it would pass in comparisons down
the road. Instead of adding the CNVs to the final results in the loop,
all CNVs for a chromosome are added only when the whole chromosome has
been checked.

I should add some more tests though to make sure that this doesn't break
in other cases.

* Refactor CNV merging to make testing simpler

In addition, .tests/units was renamed to .tests/unit, and the test
itself was made into a table test according to the guidelines in the
hydra documentation.

* Fix potential issue if skip would have been None

* Add test case from real report

* Update path for pytest github action