feat: add jumble to reports module (#80)
* feat: add jumble to common.smk

* feat: add jumble parsing in cnv_json.py

* feat: add jumble lable to merge_cnv_json

* feat: add jumble to test config

* test: add jumble

* test: add jumble

* test: add jumble test files

* fix: handle chromosomes with missing chr

* fix: add chr to chrom in jumble output

* fix: handle chromosomes with missing chr

* fix: bugfix

* Update jumble test files

This strips the `chr` prefix from the chromosome names, since this is
what Jumble does. In addition to this, the log2 ratios and segments are
updated to not be identical to those from cnvkit. This will help spot
certain issues.

* fix: more robust data keys

* style: prettier formatting

* Update chromosome names for Jumble variants

---------

Co-authored-by: Niklas Mähler <2573608+maehler@users.noreply.github.com>

fix: rendering issue of summarised log2 ratios when changing caller