feat(cnv): ability to manually adjust tumor cell content in plots (#96)
* Add TC controls

Logic to be added

* Add control for turning adjustment on/off

* Implement log2 adjustment in chromosome plot

* Scale VAF with purity

* Simpler computation of ratio adjustment

* Scale log2 and vaf in genome plot

There are still some issues with the summarised bars sticking out beyond
the plot boundaries, but this should be solvable with a clip rect (that
I thought I had already implemented). Animations also have to be added
for consistency.

* Fix clip-path for VAF panels

* Fix transitions for log2 ratios in genome view

* Fix transitions for VAF in genome view

* Same transitions for ratios in chromosome and genome view

* Same transitions for VAF in chromosome and genome view

* Set lower TC threshold to 0.05

This to avoid undefined values in the calculations.

* Mark summary outliers with triangles

Outliers in this case are points within a window that hit the minimum
copy number value. If this happens for any point within a window, the
marker is shown at the bottom of the plot, regardless of the scale.

* Add outlier triangles to genome plot

* Better outlier handling in chromosome plot

* Fix bug with disappearing VAFs

If the VAFs became zero due to modified TC, they would not show up again
when stepping the TC up to a level where they should be non-zero. This
was due to a mistake in the key function where zero-values would be
misinterpreted.