feat(cnv): better help section (#100)
* Add fonttools dependency

* Add help icon

* Add modal for help text

This commit also makes the modal handling a bit more general by having a
unified handling of how the modals are closed.

* Only show plotting caveats in the help text

* Move the general help link out of the way

* Move results table text to help

Additional links to the help was also added for relevant headers, and
these will take the user to the corresponding section of the help.

* Add help text for table filtering

* Don't show table help if not included in report

* Remove help links from print layout

* Add title to close link for message modal

* Add close event listener to all dialogs

Also add preventDefault to avoid scrolling to the top of the page when
closing the help.