chore: merge develop into main (#101)
* feat(cnv): ability to manually adjust tumor cell content in plots (#96)

* Add TC controls

Logic to be added

* Add control for turning adjustment on/off

* Implement log2 adjustment in chromosome plot

* Scale VAF with purity

* Simpler computation of ratio adjustment

* Scale log2 and vaf in genome plot

There are still some issues with the summarised bars sticking out beyond
the plot boundaries, but this should be solvable with a clip rect (that
I thought I had already implemented). Animations also have to be added
for consistency.

* Fix clip-path for VAF panels

* Fix transitions for log2 ratios in genome view

* Fix transitions for VAF in genome view

* Same transitions for ratios in chromosome and genome view

* Same transitions for VAF in chromosome and genome view

* Set lower TC threshold to 0.05

This to avoid undefined values in the calculations.

* Mark summary outliers with triangles

Outliers in this case are points within a window that hit the minimum
copy number value. If this happens for any point within a window, the
marker is shown at the bottom of the plot, regardless of the scale.

* Add outlier triangles to genome plot

* Better outlier handling in chromosome plot

* Fix bug with disappearing VAFs

If the VAFs became zero due to modified TC, they would not show up again
when stepping the TC up to a level where they should be non-zero. This
was due to a mistake in the key function where zero-values would be
misinterpreted.

* feat(cnv): better help section (#100)

* Add fonttools dependency

* Add help icon

* Add modal for help text

This commit also makes the modal handling a bit more general by having a
unified handling of how the modals are closed.

* Only show plotting caveats in the help text

* Move the general help link out of the way

* Move results table text to help

Additional links to the help was also added for relevant headers, and
these will take the user to the corresponding section of the help.

* Add help text for table filtering

* Don't show table help if not included in report

* Remove help links from print layout

* Add title to close link for message modal

* Add close event listener to all dialogs

Also add preventDefault to avoid scrolling to the top of the page when
closing the help.