fix(cnv): bug with missing results in table (#115)
* Add failing unit test

* Add duplicate to test

* Fix bug

I'm stupid... How could I not have spotted this before? I think it was
due to assuming that the index combination on the resulting dict would
result in unique assignments, but then I didn't account for the
different sets of variants. Again, I'm stupid.

* refactor: use cnv class throughout and break out merging

* Fix bug with empty table

At some point when updating the handling of the table columns I forgot
about the case with an empty table. Now this should be addressed. Not
the most elegant solution, but it works.

* Sanity check for the number of input vcf files

* Update test according to previous refactor

* Correct test documentation

* Fix cnv equality check

Since we're now working with the CNV class, the equality must disregard
the filter status for things to work.

* Add multiple vcfs for integration test

* Add merge test based on integration test files

* Sort cnvs by chromosome and start

* Python 3.8 compatibility

* Account for mix of str and int in cnv sorting